Log: fencing: add fencer call option functions
... instead of set_bit()/clear_bit(), which improves trace messages like:

    Bit 0x00000040 set by some_function:100

to:

    Fencer call flags 0x00000040 (st_opt_topology) for some_caller set by some_function:100

+++ daemons/fenced/pacemaker-fenced.h
@@ -118,8 +118,8 @@ typedef struct remote_fencing_op_s {
     char *delegate;
     /*! The point at which the remote operation completed */
     time_t completed;
-    /*! The stonith_call_options associated with this remote operation */
-    long long call_options;
+    //! Group of enum stonith_call_options associated with this operation
+    uint32_t call_options;
 
     /*! The current state of the remote operation. This indicates
      * what stage the op is in, query, exec, done, duplicate, failed. */
